BIG GUY
REFACTOR HOW STATIC LINKING WORKS:
link game to RSDK, set `RETRO_STANDALONE=0` for RSDK
add game to RSDK include path (uses new GameMain.h)

ADDED PORTAUDIO DEVICE
this should be favored over SDL

ADDED PROPER CMAKE COMPILATION
windows (DX9 and OGL) and linux (OGL) have been tested
windows favors vcpkg, linux favors pkgconfig
both will use sources for most if packages are missing
to compile with mania, use the new cmake in mania with -DWITH_RSDK=1 (defaulted to on, turn off with 0 or N or wtv)

(READMEs will be updated soon as cmake is now the preferred method of configuring)

NEW RETRO_DEV_EXTRA DEFINE
dev menu, empty spot up top

many many small bug fixes (especially to the function table)
many small edits done to help out some compilers

+++ RSDKv5/RSDK/Graphics/Animation.hpp
@@ -79,8 +79,8 @@ struct Animator {
 
 extern SpriteAnimation spriteAnimationList[SPRFILE_COUNT];
 
-uint16 LoadSpriteAnimation(const char *filename, int32 scope);
-uint16 CreateSpriteAnimation(const char *filename, uint32 frameCount, uint32 animCount, int32 scope);
+uint16 LoadSpriteAnimation(const char *filename, uint8 scope);
+uint16 CreateSpriteAnimation(const char *filename, uint32 frameCount, uint32 animCount, uint8 scope);
 
 inline uint16 FindSpriteAnimation(uint16 aniFrames, const char *name)
 {
